@@ -16,17 +16,17 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "illuminate/support": "5.8.*|^6.0|^7.0|^8.0"
+        "php": "^8.0",
+        "illuminate/support": "^9.0"
     },
     "require-dev": {
-        "laravel/legacy-factories": "^1.0.4",
-        "orchestra/testbench": "^6.0",
-        "phpunit/phpunit": "^9.0"
+        "orchestra/testbench": "^7.0",
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
-            "Hamidrezaniazi\\Upolo\\": "src"
+            "Hamidrezaniazi\\Upolo\\": "src",
+            "Hamidrezaniazi\\Upolo\\Database\\Factories\\": "database/factories"
         }
     },
     "autoload-dev": {
