@@ -1,45 +1,52 @@
 <?php
 
-/* @var $factory Factory */
+namespace Hamidrezaniazi\Upolo\Database\Factories;
 
-use Faker\Generator as Faker;
 use Hamidrezaniazi\Upolo\Guard;
-use Illuminate\Database\Eloquent\Factory;
+use Hamidrezaniazi\Upolo\Models\File;
+use Hamidrezaniazi\Upolo\Tests\Models\MockModel;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+/**
+ * @extends Factory<File>
+ */
+class FileFactory extends Factory
+{
+    protected $model = File::class;
 
-$factory->define(\Hamidrezaniazi\Upolo\Models\File::class, function (Faker $faker) {
-    return [
-        'disk'       => 'public',
-        'uuid'       => $faker->uuid,
-        'path'       => 'public/file.png',
-        'filename'   => $faker->unique()->word,
-        'type'       => $faker->unique()->word,
-        'mime'       => $faker->unique()->word,
-        'creator_id' => factory(Guard::getGuardClassName()),
-    ];
-});
+    public function definition()
+    {
+        $creator = Guard::getGuardClassName();
 
-$factory->state(\Hamidrezaniazi\Upolo\Models\File::class, 'has_owner', function () {
-    $owner = \factory(\Hamidrezaniazi\Upolo\Tests\Models\MockModel::class)->create();
+        return [
+            'disk'       => 'public',
+            'uuid'       => $this->faker->uuid,
+            'path'       => 'public/file.png',
+            'filename'   => $this->faker->unique()->word,
+            'type'       => $this->faker->unique()->word,
+            'mime'       => $this->faker->unique()->word,
+            'creator_id' => $creator::factory(),
+        ];
+    }
 
-    return [
-        'owner_id'   => $owner->getKey(),
-        'owner_type' => $owner->getMorphClass(),
-    ];
-});
+    public function hasOwner(): self
+    {
+        return $this->state(function (array $attributes) {
+            $owner = MockModel::factory()->create();
 
-$factory->state(\Hamidrezaniazi\Upolo\Models\File::class, 'has_flag', function (Faker $faker) {
-    return [
-        'flag' => $faker->unique()->word,
-    ];
-});
+            return [
+                'owner_id'   => $owner->getKey(),
+                'owner_type' => $owner->getMorphClass(),
+            ];
+        });
+    }
+
+    public function hasFlag(): self
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'flag' => $this->faker->unique()->word,
+            ];
+        });
+    }
+}
